% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_normal.R
\name{cdf_normal}
\alias{cdf_normal}
\title{cdf_normal}
\usage{
cdf_normal(x, mean = 0, stddev = 1)
}
\arguments{
\item{x}{a double.}

\item{mean}{a double is the mean ( crest ).}

\item{stddev}{a double the standard deviation.}
}
\value{
a double
}
\description{
the cumulative density function ( cdf ) of normal distribution.
}
