% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_createScm.R
\name{com_createScm}
\alias{com_createScm}
\title{com_createScm}
\usage{
com_createScm(com)
}
\arguments{
\item{com}{the Conditional Occurrence Matrix.}
}
\value{
a data.frame with probabilities. a stochastic matrix.
}
\description{
create a scm (StoChastic Matrix) from a com (Conditional Occurrence Matrix).
basicaly what this will do id to divide each value for the sum of the values of their row.
}
\references{
Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, chapter: 3.
}
