% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_erf.R
\name{mat_erf}
\alias{mat_erf}
\title{mat_erf}
\usage{
mat_erf(x)
}
\arguments{
\item{x}{a double for the function.}
}
\value{
a double.
}
\description{
the erf ( ERror Function ) also called ( gauss error function ).
}
