% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_erf.R
\name{math_erf}
\alias{math_erf}
\title{math_erf}
\usage{
math_erf(x)
}
\arguments{
\item{x}{a double for the function.}
}
\value{
a double.
}
\description{
the erf ( ERror Function ) also called ( gauss error function ).
}
