% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_createMultivariateFromExogCom.R
\name{matrix_createMultivariateFromExogCom}
\alias{matrix_createMultivariateFromExogCom}
\title{matrix_createMultivariateFromExogCom}
\usage{
matrix_createMultivariateFromExogCom(endog, exog, tPlusX = 1L)
}
\arguments{
\item{endog}{a factor vector.}

\item{exog}{a factor vector, the levels need to be the same as in endog.}

\item{tPlusX}{the order ( lag ) for the function. 1 means: does the previous value of exog
predicts the actual value of endog.}
}
\value{
a data.frame with integers, rownames are the levels of exog and columns are the levels of endog.
}
\description{
this will calculate the conditional number of occorrences from levels of a sequence (exog)
cause another one (endog) C(endog|exog) number occurrences of endog given exog.
}
