% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_createMultivariateMultipleFromExogsCom.R
\name{matrix_createMultivariateMultipleFromExogsCom}
\alias{matrix_createMultivariateMultipleFromExogsCom}
\title{matrix_createMultivariateMultipleFromExogsCom}
\usage{
matrix_createMultivariateMultipleFromExogsCom(endog, exogs, tPlusX = 1L)
}
\arguments{
\item{endog}{a factor vector, is the endog (cause).}

\item{exogs}{a list with factors vectors, the (givens).}

\item{tPlusX}{a integer is the leading to be applied to the endog 1 means: ('actual exogs can bredics the next endog').}
}
\value{
a data.frame with number of occurrences.
}
\description{
create a conditional occurrence matrix (com) from multiple exogs, this should
to accept one 'endog' factor vector and multiple 'exogs' factors vectors. the com returned
is the com of states of the exogs happening together causes endog.
}
