% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_discretize.R
\name{matrix_discretize}
\alias{matrix_discretize}
\title{matrix_discretize}
\usage{
matrix_discretize(data, n, n.custom = NULL)
}
\arguments{
\item{data}{a data.frame.}

\item{n}{a integer with the maximum number of discrete values.}

\item{n.custom}{takes a list with the name of the column and the number of chunk.}
}
\description{
transform a data.frame with continuous values into a data.frame
with 'x<=VALUE' value, only numeric columns will be transformed.
}
