% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_multiplication.R
\name{matrix_multiplication}
\alias{matrix_multiplication}
\title{matrix_multiplication}
\usage{
matrix_multiplication(matrix1, matrix2)
}
\arguments{
\item{matrix1}{the first matrix}

\item{matrix2}{the second matrix}
}
\value{
a data.frame of multiplied matrices
}
\description{
do multiplication of two matrices.
}
\references{
https://online.stat.psu.edu/statprogram/reviews/matrix-algebra/arithmetic#:~:text=To\%20perform\%20matrix\%20multiplication\%2C\%20the,columns\%20of\%20the\%20second\%20matrix.
https://en.wikipedia.org/wiki/Matrix_multiplication
https://matrix.reshish.com/multCalculation.php
Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, pags.: 103-104.
}
