% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_power.R
\name{matrix_power}
\alias{matrix_power}
\title{matrix_power}
\usage{
matrix_power(.matrix, n)
}
\arguments{
\item{.matrix}{the matrix to be 'powered'}

\item{n}{the exponential.}
}
\value{
a data.frame with the powered matrix.
}
\description{
do 'power' (exponentiation) of matrix
}
\references{
#' Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, pags.: 103-104.
}
