% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_metropolisHastings.R
\name{mcmc_metropolisHastings}
\alias{mcmc_metropolisHastings}
\title{mcmc_metropolisHastings}
\usage{
mcmc_metropolisHastings(scm, x_0, n, dist = scm_steadyReady(scm))
}
\arguments{
\item{scm}{a data.frame with a stochastic matrix.}

\item{x_0}{a character the initial state.}

\item{n}{the horizon (num steps) of the sampling.}

\item{dist}{the stable distribution (pi function).}
}
\value{
a character vector containing the results of sampling
}
\description{
the metropolis-hastings algorithm, a
improvement for metropolis algorithm (metropolis: 1953; hastings: 1970).
}
\section{expected comportment}{

is expected in low samples the distribution of the results are different than the stable distribution (steadyReady)
and for long long run the distribution of the results are near equal the stable distribution.
}

\references{
https://en.wikipedia.org/wiki/Metropolis–Hastings_algorithm
https://stephens999.github.io/fiveMinuteStats/MH_intro.html
}
