% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mc_MTD.R
\name{model_mc_MTD}
\alias{model_mc_MTD}
\title{model_mc_MTD}
\usage{
model_mc_MTD(data, l, t = length(data))
}
\arguments{
\item{data}{a vector of factors.}

\item{l}{the order.}

\item{t}{is the actual time, (dont need to be changed).}
}
\description{
Mixture Transition Distribution model. a model for high order markov chain.
}
\references{
The Mixture Transition Distribution Model for High-Order Markov Chains and Non-Gaussian Time Series.
André Berchtold and Adrian E. Raftery
Statistical Science 2002, Vol. 17, No. 3, 328–356
}
