% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mmc_ching2008.R
\name{model_mmc_ching2008}
\alias{model_mmc_ching2008}
\title{model_mmc_ching2008}
\usage{
model_mmc_ching2008(
  data,
  n,
  norm.type = "1",
  s = ncol(data),
  j = 1:s,
  r = nrow(data) - 1,
  m = length(levels(data[[1]])),
  options.mco.popsize = s * 4,
  options.mco.generations = 100
)
}
\arguments{
\item{n}{is the order.}

\item{norm.type}{is the type of norm, it can be any type in the base::norm, from the paper ther recommendations is:
'1': (norm-1 (taxicab norm))always give the most robust answer.
'I': (norm-infinity) avoids gross discrepancies with the data.
'2': (norm-euclidian) if the errors are known to be normally distributed then is the best choice.}

\item{s}{is the number of sequences.}

\item{j}{is a iterator (vector) for the number of sequences (s).}

\item{r}{is the previous time in the data. (t-1).}

\item{m}{is the number of levels.}
}
\description{
a multivariate markov chains model from chin et all 2008.
}
\references{
Higher-order multivariate Markov chains and their applications
Wai-Ki Ching, Michael K. Ng, Eric S. Fung
doi: 10.1016/j.laa.2007.05.021
}
