% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mmc_FMMC.R
\name{model_mmc_FMMC}
\alias{model_mmc_FMMC}
\title{model_mmc_FMMC}
\usage{
model_mmc_FMMC(
  data,
  k,
  r,
  n = nrow(data),
  m = ncol(data),
  s = length(levels(data[[k]])),
  options = list(generations = 100, n.threads = min(c(parallel::detectCores(), r)))
)
}
\arguments{
\item{data}{a data.frame containing columns, each data.frame needs to have the sames factors levels.}

\item{k}{the index of the endog column.}

\item{r}{is the order or the model. if how much past data probabilities to consider.}

\item{n}{is the actual time in the data. (taken automaticaly).}

\item{m}{is the number of datas. (taken automaticaly).}

\item{s}{is the number of states. (taken automaticaly).}

\item{options}{takes a list with values for additional options, these options there'nt no correlation with the model
itself but are paramether for things used to implement, for example the n.threads.}
}
\value{
return a numeric vector with the probabilities for X_n+1 (the next state). the names of the vector
are the names of the levels of factor.
}
\description{
implements the model FM-MC from their paper.
}
\references{
A Gradual Facilitate High‑Order Multivariate Markov Chains
Model with Application to the Changes of Exchange Rates
in Egypt: New Approach
A. M. Elshehawey
https://doi.org/10.1007/s42519-021-00179-y
}
