% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mmc_MTDProbit.R
\name{model_mmc_MTDProbit}
\alias{model_mmc_MTDProbit}
\title{model_mmc_MTDProbit}
\usage{
model_mmc_MTDProbit(
  data,
  j,
  l,
  nj0 = 0,
  fiFunc = function(d) {
     cdf_normal(d, mean = 0, stddev = 1)
 },
  s = ncol(data),
  t = nrow(data),
  options = list()
)
}
\arguments{
\item{data}{a data.frame with sequences.}

\item{j}{the number of the column in the data is the endog (who will be returned the probabilities).}

\item{l}{the number of order.}

\item{nj0}{a initial value (as said in the paper it can be zero: a constant term nj0 is introduced in the Pfij
specification, and in this way, the proposed specification involves one additional parameter in comparison with the
MTD case; although it can be set to zero, nj0 generally improves the fit (i.e. allows the probability Pfij to be
closer to Pj).}

\item{fiFunc}{can be any probability function, the default is the cumulative distribution function of a standard normal
distribution. (cdf of normal function with mean=0 and stddev=1). but as said in the paper: fi can be replaced by
another distribution function of any continuous random variable with state space of Real.}

\item{s}{the number of sequences.}

\item{t}{the actual time.}

\item{options}{ia a list with parameters for the model, these parameters are:
\itemize{
\item{opt.opts: }{parameters to be passed to the parameter 'opts' of the function nloptr::nloptr}
}}
}
\value{
a numeric vector with values (probabilities).
}
\description{
implementation of a multivariate markov chains MTD-Probit model from their paper.
}
\references{
A New Model for Multivariate Markov Chains
JOÃO NICOLAU
doi: 10.1111/sjos.12087
}
