% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdf_normal}
\alias{pdf_normal}
\title{Rcpp_pdf_normal}
\usage{
pdf_normal(x, mean, stddev)
}
\arguments{
\item{x}{the target probability}

\item{mean}{the mean of the distribution}

\item{stddev}{the standard deviation}
}
\value{
a double with PDF
}
\description{
normal distribution PDF from: "https://web.stanford.edu/class/archive/cs/cs109/cs109.1178/lectureHandouts/110-normal-distribution.pdf".
}
