% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distDiscrete.R
\name{sample_distDiscrete}
\alias{sample_distDiscrete}
\title{sample_distDiscrete}
\usage{
sample_distDiscrete(n = 1, dist)
}
\arguments{
\item{n}{a integer with the number of samples.}

\item{dist}{a numeric vector with names, names contains the results
and the values contains the probabilities}
}
\value{
a character vector.
}
\description{
take a sample from a discrete distribution.
}
