% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distUniform.R
\name{sample_distUniform}
\alias{sample_distUniform}
\title{sample_distUniform}
\usage{
sample_distUniform(n = 1, min = 0, max = 1)
}
\arguments{
\item{n}{means "numbers" is the number of samples to get.}

\item{min}{the minium return value.}

\item{max}{the maximum return value.}
}
\value{
a double.
}
\description{
get a sample from a uniform distribution
}
