% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_fixEigenvectorLowerThanOne.R
\name{scm_fixEigenvectorLowerThanOne}
\alias{scm_fixEigenvectorLowerThanOne}
\title{scm_fixEigenvectorLowerThanOne}
\usage{
scm_fixEigenvectorLowerThanOne(scm, maximumDifference = 0.02)
}
\arguments{
\item{scm}{a data.frame with a stochastic matrix}

\item{maximumDifference}{the maximum difference from one to accept to fix these values.
will raise an error if the difference is highter than this param value.}
}
\value{
a scm (stochastic matrix) with fixed values now the rum of each
row will be equal to one.
}
\description{
will fix values that where their eigenvector (sum of all values in a row) is
lower than one, this can happen for example in the division, the division
by a number can return a repeating decimal (primes are very common).
Rows with eigenvector that dont end on exactly one will be
fixed using the formula:
lowerstValueOfRow= 1-(A-L)
A: sum of all values in the row
L: the lowerst value in the row.
}
