% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_multiply.R
\name{scm_multiply}
\alias{scm_multiply}
\title{scm_multiply}
\usage{
scm_multiply(scm1, scm2)
}
\arguments{
\item{scm1}{data.frame the first scm}

\item{scm2}{data.frame the second scm}
}
\value{
a multiplied stochastic matrix
}
\description{
do matrix multiplication
}
\references{
https://www.khanacademy.org/math/precalculus/x9e81a4f98389efdf:matrices/x9e81a4f98389efdf:multiplying-matrices-by-matrices/v/matrix-multiplication-intro
}
