% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_nStepTransitionProbability.R
\name{scm_nStepTransitionProbability}
\alias{scm_nStepTransitionProbability}
\title{scm_nStepTransitionProbability}
\usage{
scm_nStepTransitionProbability(scm, state, n)
}
\arguments{
\item{scm}{the stochastic matrix with probabilities}

\item{state}{the initial probabilities}

\item{n}{the horizon}
}
\description{
calculate the probabilities n-step ahead given the probabilities
using a stochastic matrix.
}
\references{
#' Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, chapter: 4.
}
