% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_statesTimes.R
\name{scm_statesTimes}
\alias{scm_statesTimes}
\title{scm_statesTimes}
\usage{
scm_statesTimes(scm)
}
\arguments{
\item{scm}{a data.frame with a right stochastic matrix}
}
\value{
a numeric vector with names and values with the percentage of time spent on each state.
}
\description{
calculate the average time (percent) the system will stay at state.
}
\references{
Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, chapter: 7.3.
}
