% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_steadyState.R
\name{scm_steadyState}
\alias{scm_steadyState}
\title{scm_steadyState}
\usage{
scm_steadyState(scm, n.max = 1000)
}
\arguments{
\item{scm}{a data.frame with the stochastic matrix}

\item{n.max}{a integer the maximum number of steps for simulating.}
}
\value{
a numeric vector with names and probabilities for each state.
}
\description{
find the steadyState probabilities (stable distribution) for
a right stochastic matrix.
}
\references{
Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, chapter: 4.
}
