% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_transitionProbability.R
\name{scm_transitionProbability}
\alias{scm_transitionProbability}
\title{scm_transitionProbability}
\usage{
scm_transitionProbability(scm, state)
}
\arguments{
\item{scm}{the stochastic matrix.}

\item{state}{a numeric vector containing names with the priori probabilities.}
}
\value{
a numeric vector containing the probabilities
}
\description{
calculate the one step transition probability.
}
\references{
Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, chapter: 4.
}
