% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_verifyScm.R
\name{scm_verifyScm}
\alias{scm_verifyScm}
\title{scm_verifyScm}
\usage{
scm_verifyScm(scm)
}
\arguments{
\item{scm}{a data.frame with a stochastic matrix.}
}
\value{
the same input. but if the scm is not valid will raise an error.
}
\description{
this function verifies if the scm is ready for usage.
}
\references{
Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, chapter: 3.
}
