% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slw_appleChoice.R
\name{slw_appleChoice}
\alias{slw_appleChoice}
\title{slw_appleChoice}
\usage{
slw_appleChoice(dataAll, current, states, labelColumn, tPlusX)
}
\arguments{
\item{dataAll}{a data.frame containing all datas.}

\item{current}{a data.frame with one row containing actual row data (the colnames of this data.frame needs to be qual
to the dataAll data.frame).}

\item{states}{a character (string) containing the states (following the rownames standard in shinnigamiLeftWing)}

\item{labelColumn}{the name of the column containing the labels.}

\item{tPlusX}{the predict horizon (as in shinnigamiLeftWing).}
}
\value{
a list with the states names and a duplicates numeric containing the number of duplicates in each label state.
}
\description{
counts the similatiries between a row of a data.frame and rows of a data.frame which that states
happened (current).
}
