% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slw_battleForTheApple.R
\name{slw_battleForTheApple}
\alias{slw_battleForTheApple}
\title{slw_battleForTheApple}
\usage{
slw_battleForTheApple(
  dataAll,
  data1,
  data2,
  tPlusX,
  labelColumn = "label",
  minOccurrences = 3,
  maxDefensors = 10,
  maxAttackers = 10
)
}
\arguments{
\item{dataAll}{a data.frame containing the raw states data which the function will search in using data1 and data2 rownames.}

\item{data1}{data.frame with one column (you can do it using drop=FALSE in a data.frame selector)}

\item{data2}{data.frame with one column (you can do it using drop=FALSE in a data.frame selector)}

\item{tPlusX}{the distance in the future to predict, THIS MUST BE THE SAME WAS USED IN THE shinnigamiLeftWing function.}

\item{labelColumn}{a character is the name of the column of 'dataAll' which contains the answer.}

\item{minOccurrences}{the minimum number of occurrences to stop, for example: if two states not occurs together or
occurs with a low number of times the function will try with the next value of the 'alternative hipoteses' against the
'null hipoteses', always will be 'alternative vs null' not the inverse. at end of max number of attackers
the functil will change the defensor (rowname of null hipoteses) to the next value and will reset the index of the attacker
(alternative hipotesis). \cr
SET IT TO A HIGH VALUE OR INF TO DISABLE IT and run for all values in data1 and data2.}

\item{maxDefensors}{a integer is the maximum number of values in the 'null hipoteses'. \cr
SET IT TO A HIGH VALUE OR INF TO DISABLE IT.}

\item{maxAttackers}{a integer is the maximum number of values in the 'alternative hipoteses'. \cr
SET IT TO A HIGH VALUE OR INF TO DISABLE IT.}
}
\value{
a data.frame with these columns: {column of data1} {column of data2} defensor.index attacker.index \cr
defensor.index: is the 'rank' of the data.frame containing the probability of the null hipotesis (highest probability). \cr
attacker.index: is the 'rank' of the data.frame containing the probability of the alternative hipotesis. \cr
for example: if the defensor.index=1 this means that the defensor will be the value of the highest probability in the
null hipotesis, and if this value is: defensor.index=2 this means the value used is the secong highest. the same with
the 'attacker.index'.
}
\description{
do the states of the output of the function shinnigamiLeftWing (slw) to battle the actual
hipoteses (null hipoteses) vs a oposite hipoteses (alternative hipoteses).
is to solve the problem of the outputs of shinnigamiLeftWing function.
}
\section{problem}{

the shinnigamiLeftWing has a problem which two near equal probabilities in their results are near
opose equaly, for example: you have two possible columns 'fell' and 'rose', so in the results of the function you get
something near, 0.9 fall and 0.89 rose, so the question is: "how to solve this ploblem".
}

\section{solution}{

the solution which this function tries to apply is to get these states which occurer with high opose simultaneous
probability and check in the complete data when this happened in the past. for example: if the slw says that when
A=2 & B=5: 0.9 fell 0.1 rose, and, C=1 & D=3: 0.11 fell 0.89 rose, just to check when these states happened together
is to get the probability of: A=2 & B=5 & C=1 & D=3, with it you will have a level 4 result, with you can use
it as a tiebreaker.
}

\section{rite}{

this function will do these steps in their work to solve the slw problem: 1 - get the null and alternative hipoteses,
the null hipoteses will be the column which has the maximum value (maximum probability) and the alternative is the
remaining, them are data.frames with one column; 2 - order both the data.frames (null hipoteses and alternative hipote
sis) decreasingment; 3 - get in 'dataAll' the rows which bot states happen together and count of these occurrences
which was the result them.
}

