% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slw_happened.R
\name{slw_happened}
\alias{slw_happened}
\title{slw_happened}
\usage{
slw_happened(current, states, levels)
}
\arguments{
\item{current}{a data.frame with one row.}

\item{states}{a character vector containing at least one character.}

\item{levels}{a integer vector containing the levels to return. this should to be the same level than the states cause
if for example, the states contains characters of level3 but you choose level1, so any value with any column like
in that state will be returned.\cr
example: states= "exog1=2 & exog2=5 & exog3=2", levels=1; in this case a row which contains only exog1=2... is a valid.}
}
\value{
the positions in 'states' vector which happens in 'current'.
}
\description{
check if/which rownames states (shinnigamiLeftWing rowname) happened in the current data.frame
(a data.frame with one row)
}
