% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slw_stateToVector.R
\name{slw_stateToVector}
\alias{slw_stateToVector}
\title{slw_stateToVector}
\usage{
slw_stateToVector(data)
}
\arguments{
\item{data}{the state character}
}
\value{
a character vector
}
\description{
tranform a state standard of shinnigamiLeftWing to a named vector of character
}
