% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slw_whichRowsHappens.R
\name{slw_whichRowsHappens}
\alias{slw_whichRowsHappens}
\title{slw_whichRowsHappens}
\usage{
slw_whichRowsHappens(dataAll, states)
}
\arguments{
\item{dataAll}{the complete data.frame}

\item{states}{a character with the pattern of shinnigamiLeftWing rownames}
}
\value{
a numeric vector
}
\description{
returns the number of rows which states happened.
}
