% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_countNOccurrencesOfEachRowString.R
\name{util_countNOccurrencesOfEachRowString}
\alias{util_countNOccurrencesOfEachRowString}
\title{util_countNOccurrencesOfEachRowString}
\usage{
util_countNOccurrencesOfEachRowString(uniques, data)
}
\arguments{
\item{uniques:}{a data.frame with uniques values to be counted.}

\item{data:}{the complete data.frame where to count.}
}
\value{
a numeric vecto these values are the number of occurrences for each row in uniques
happened in data.
}
\description{
count how much times a row of string happened in the given data.frame.
this work only with strings (character).
}
