% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_createCom.R
\name{vector_createCom}
\alias{vector_createCom}
\title{vector_createCom}
\usage{
vector_createCom(vec, tPlusx = 1L)
}
\arguments{
\item{vec}{a vector of factors.}

\item{tPlusx}{the difference of step of the next value. is globaly used t_n+1.}
}
\value{
a data.frame with equal number of rows and columns and equal names in colnames and rownames with
number of occurences.
}
\description{
this function creates a matrix with the number of conditional occurrences. the main purpose
of the usage of this function is to create a stochastic matrix (markov chain matrix).
'Com' means: Conditional Occurrence matrix.
}
\references{
Markov Chains, From Theory to Implementation and Experimentation, Paul A. Gagniuc, chapter: 3.
}
