% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_discretize.R
\name{vector_discretize}
\alias{vector_discretize}
\title{vector_discretize}
\usage{
vector_discretize(data, mode = "all", options = list())
}
\arguments{
\item{data}{a numeric vector.}

\item{mode}{how to do the discretization? can be: "all","n.factors","fixed.minMaxStep"}

\item{options}{a list with the param for that mode. each mode there itself parameters.}
}
\value{
a factor vector.
}
\description{
transform a numeric vector into a discrete
}
\section{mode= "all"}{

is the default mode it just transform the data into factor return(as.factor(data))
this therent params.
}

\section{mode= "n.factors"}{

split the data into numbers of factors.
n: a integer is how much factors to split
}

\section{mode= "fixed.minMaxStep"}{

equaly spaced factors, from a min to a max each step. this will use the 'seq' function to generate a sequence
which these parameters.
the factors of this mode will be: [min, max).
min: a double is the minimum value.
max: a double is the maximum value.
step: a double is the 'step' for the sequence, this will be the 'by' of the function 'seq'.
}

