% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_occurrencyProbability.R
\name{vector_occurrencyProbability}
\alias{vector_occurrencyProbability}
\title{vector_occurrencyProbability}
\usage{
vector_occurrencyProbability(vec)
}
\value{
a named numeric vector with probabilities
}
\description{
a very simple probability of vector to came that value (factor).
it is basicaly nOccurrencyOfThatFactor/lengthOfVector.
}
