% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_splitIntoChunks.R
\name{vector_splitIntoChunks}
\alias{vector_splitIntoChunks}
\title{vector_splitIntoChunks}
\usage{
vector_splitIntoChunks(data, n)
}
\arguments{
\item{data}{a vector}

\item{n}{the number of chunks}
}
\value{
a list with values
}
\description{
function will split data into chunks
}
\references{
https://www.geeksforgeeks.org/split-vector-into-chunks-in-r/
}
