% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mmc_FMMC.R
\name{model_mmc_FMMC}
\alias{model_mmc_FMMC}
\title{model_mmc_FMMC}
\usage{
model_mmc_FMMC(
  data,
  k,
  r,
  n = nrow(data),
  m = ncol(data),
  s = length(levels(data[[k]]))
)
}
\description{
implements the model FM-MC from their paper.
}
\references{
A Gradual Facilitate High‑Order Multivariate Markov Chains
Model with Application to the Changes of Exchange Rates
in Egypt: New Approach
A. M. Elshehawey
https://doi.org/10.1007/s42519-021-00179-y
}
