% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mmc_MTDProbit.R
\name{model_mmc_MTDProbit}
\alias{model_mmc_MTDProbit}
\title{model_mmc_MTDProbit}
\usage{
model_mmc_MTDProbit(
  data,
  j,
  l,
  nj0,
  type,
  fiFunc = function(d) {
     cdf_normal(d, mean = 0, stddev = 1)
 },
  s = ncol(data),
  t = nrow(data)
)
}
\arguments{
\item{data}{a data.frame with factors.}

\item{j}{the index of the endog column.}

\item{l}{the number of order.}

\item{nj0}{a initial value (as said in the paper it can be zero).}

\item{type}{it can be 'dependent' or 'indepenent', dependent is if the levels of all columns in the data are equals.
'independent' is if the them are different.}

\item{fiFunc}{is a function which takes a vector with probabilities and returns a cdf (cumulative distribution function)
value for each the the default is the standard normal distribution (mean = 0, stddev = 1), but as said in the paper it can be
replaceb by any cdf that works with continum values and works with state space Reals (Real numbers, math).}

\item{s}{the the number of sequences in the data (this will only used to the exog).}

\item{t}{the present time in the data (will be used to take the previous values).}
}
\description{
a multivariate markov chains model implemented from their paper MTD-Probit.
}
\references{
A New Model for Multivariate Markov Chains
JOÃO NICOLAU
doi: 10.1111/sjos.12087

The Profitability in the FTSE 100 Index: A New Markov Chain Approach
Flavio Ivo Riedlinger, João Nicolau
doi: 10.1007/s10690-019-09282-4
}
