% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mmc_ShinnigamiLeftWing.R
\name{model_mmc_ShinnigamiLeftWing}
\alias{model_mmc_ShinnigamiLeftWing}
\title{model_mmc_ShinnigamiLeftWing}
\usage{
model_mmc_ShinnigamiLeftWing(
  endog,
  exogs,
  levels = 1,
  tPlusX = 1,
  levels.length = rep(Inf, length(levels)),
  options.nThread = parallel::detectCores(),
  options.threadType = ifelse(Sys.info()["sysname"] == "Windows", "PSOCK", "FORK")
)
}
\arguments{
\item{endog}{a factor vector containig the endog (causes).}

\item{exogs}{a data.frame with the exogs data, this is the (given) all need to be factors.}

\item{levels}{a integer vector contaning values to the levels, the min(levels) need to be >=2.}

\item{tPlusX}{the number of leading to apply to endog, tPlusX 1 means "does actual exogs predicts the next endog?".}

\item{levels.length}{the maximum of combinations to obtain from levels > 1 (level 1 will not use this). Inf if all
combinations are wanted. for example level 2 calculate only 100000 combinations and level 3 calculate 10000 you can set:
levels.length= c(100000,10000). the 100000 will be use by the first index in the levels parameter and 10000 by the
second index in the levels parameter. (as said level1 dont use this).}

\item{options.nThread}{the number of threads to run paralelly.}

\item{options.threadType}{the type of thread, on windoes this can be PSOCK but linux distros accepts FORK. see about
parallel package.}
}
\value{
a list with data.frames
}
\description{
calculate probabilities of multiple possibilities given the exogs factors.
}
\section{tip}{

after returned the results can be used the function in the package dplyr: dplyr::bind_rows to get a unique data.frame.
}

