% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_normal.R
\name{pdf_normal}
\alias{pdf_normal}
\title{pdf_normal}
\usage{
pdf_normal(x, mean = 0, stddev = 1)
}
\arguments{
\item{x}{the target}

\item{mean}{the mean of the distribution ( crest ).}

\item{stddev}{the standard deviation of the distribution.}
}
\value{
a double.
}
\description{
the normal distribution probability density function ( pdf ).
}
\references{
https://web.stanford.edu/class/archive/cs/cs109/cs109.1178/lectureHandouts/110-normal-distribution.pdf
}
